+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.2.1 - 2020-09-22
+
+### Added
+- Adds `Api::setUser()` method to enable bearer tokens in anonymous console requests  
+
+### Improved
+- The `uploaderId` attribute is now set for imported Assets  
+
 ## 1.2.0 - 2020-09-22
 
 ### Added